refactor: Resolved a service naming conflict between logical delete and physical delete operations in the user management module.
