+++ spring-ddd-domain/src/main/java/com/springddd/domain/user/WipeSysUserByIdsDomainService.java
@@ -4,7 +4,7 @@ import reactor.core.publisher.Mono;
 
 import java.util.List;
 
-public interface DeleteSysUserByIdsDomainService {
+public interface WipeSysUserByIdsDomainService {
 
     Mono<Void> deleteByIds(List<Long> ids);
 }
